@@ -2,21 +2,22 @@
     "name": "rector/phpstan-rules",
     "license": "MIT",
     "type": "phpstan-extension",
-    "description": "PHPStan rules for Rector projects - with focus on static reflection, constant re-use and Rector design patterns",
+    "description": "PHPStan rules for Rector projects - with focus on static reflection and Rector design patterns",
     "require": {
-        "php": "^7.4|^8.0",
-        "symplify/phpstan-rules": "^11.1.6",
+        "php": "^8.1",
+        "symplify/phpstan-rules": "^11.2",
         "phpstan/phpstan": "^1.8.2",
         "nette/utils": "^3.2"
     },
     "require-dev": {
-        "symplify/easy-coding-standard": "^11.1.6",
+        "symplify/easy-coding-standard": "^11.2",
         "phpstan/extension-installer": "^1.1",
         "phpunit/phpunit": "^9.5",
         "rector/rector": "dev-main",
         "phpstan/phpstan-webmozart-assert": "^1.2",
         "phpstan/phpstan-strict-rules": "^1.4",
-        "tracy/tracy": "^2.9"
+        "tracy/tracy": "^2.9",
+        "symplify/easy-ci": "^11.2"
     },
     "autoload": {
         "psr-4": {
