@@ -30,11 +30,9 @@ final class NoInstanceOfStaticReflectionRule extends AbstractSymplifyRule
      */
     public const ERROR_MESSAGE = 'Instead of "instanceof/is_a()" use ReflectionProvider service or "(new ObjectType(<desired_type>))->isSuperTypeOf(<element_type>)" for static reflection to work';
 
-    private AllowedAutoloadedTypeAnalyzer $allowedAutoloadedTypeAnalyzer;
-
-    public function __construct(AllowedAutoloadedTypeAnalyzer $allowedAutoloadedTypeAnalyzer)
-    {
-        $this->allowedAutoloadedTypeAnalyzer = $allowedAutoloadedTypeAnalyzer;
+    public function __construct(
+        private readonly AllowedAutoloadedTypeAnalyzer $allowedAutoloadedTypeAnalyzer
+    ) {
     }
 
     /**
@@ -81,10 +79,7 @@ CODE_SAMPLE
         ]);
     }
 
-    /**
-     * @param FuncCall|Instanceof_ $node
-     */
-    private function resolveExprStaticType($node, Scope $scope): ?Type
+    private function resolveExprStaticType(FuncCall|Instanceof_ $node, Scope $scope): ?Type
     {
         if ($node instanceof Instanceof_) {
             return $this->resolveInstanceOfType($node, $scope);
