@@ -45,7 +45,7 @@ final class PhpUpgradeImplementsMinPhpVersionInterfaceRule implements Rule
     {
         /** @var string $className */
         $className = (string) $node->namespacedName;
-        if (substr_compare($className, 'Rector', -strlen('Rector')) !== 0) {
+        if (! str_ends_with($className, 'Rector')) {
             return [];
         }
 
