@@ -4,23 +4,20 @@ declare(strict_types=1);
 
 namespace Rector\PHPStanRules\Tests\ReturnTypeExtension\NameResolverReturnTypeExtension;
 
+use Iterator;
 use PHPStan\Testing\TypeInferenceTestCase;
 
 final class NameResolverReturnTypeExtensionTest extends TypeInferenceTestCase
 {
-    /**
-     * @return iterable<string, mixed[]>
-     */
-    public function dataAsserts(): iterable
+    public function dataAsserts(): Iterator
     {
         yield from $this->gatherAssertTypes(__DIR__ . '/data/get_name_class_method.php.inc');
     }
 
     /**
      * @dataProvider dataAsserts()
-     * @param mixed ...$args
      */
-    public function testAsserts(string $assertType, string $file, ...$args): void
+    public function testAsserts(string $assertType, string $file, mixed ...$args): void
     {
         $this->assertFileAsserts($assertType, $file, ...$args);
     }
